requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.3
django-extensions==3.2.3
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
djangorestframework==3.13.1